Rename and add missing Talent Events
