+++ Communication/Packets/Incoming/Talent/GuideAdvertisementReadEvent.cs
@@ -1,0 +1,8 @@
+﻿using Plus.HabboHotel.GameClients;
+
+namespace Plus.Communication.Packets.Incoming.Talent;
+
+internal class GuideAdvertisementReadEvent
+{
+    public Task Parse(GameClient session) => throw new NotImplementedException();
+}
